refactor(agent): move AgentConversationHarness under fixtures/helpers
The class coordinates actions across the app without owning a locator
surface, which is what fixtures/helpers holds; the root fixture file now
carries only the fixture registration.

Addresses review feedback:
https://github.com/Comfy-Org/ComfyUI_frontend/pull/16649#discussion_r3917334729
