+++ browser_tests/fixtures/agentConversationFixture.ts
@@ -1,369 +1,6 @@
-import type { Locator, Page, WebSocketRoute } from '@playwright/test'
-import { expect } from '@playwright/test'
-
-import { applyOps, mint } from '@comfyorg/comfy-multi-player'
-import type { WidgetCatalog, WorkflowJSON } from '@comfyorg/comfy-multi-player'
-import * as Y from 'yjs'
-
-import enMessages from '@/locales/en/main.json' with { type: 'json' }
-import type { GraphOperation } from '@/workbench/extensions/agent/crdt/graphOperations'
-import { mintWireOps } from '@/workbench/extensions/agent/crdt/opEnvelope'
-import type {
-  AgentCancelAccepted,
-  AgentMessages,
-  AgentTurnAccepted,
-  AgentWsEvent
-} from '@/workbench/extensions/agent/schemas/agentApiSchema'
-import { parseAgentWsEvent } from '@/workbench/extensions/agent/schemas/agentApiSchema'
-
-import { agentTest, bootAgentApp } from '@e2e/fixtures/agentPanelFixture'
-import type {
-  AgentConversation,
-  RecordedWsEvent
-} from '@e2e/fixtures/data/agent/agentConversation'
+import { agentTest } from '@e2e/fixtures/agentPanelFixture'
 import { loadAgentConversation } from '@e2e/fixtures/data/agent/agentConversation'
-import { jsonRoute } from '@e2e/fixtures/utils/jsonRoute'
-
-const THREAD_ID = 'e9a2f3d1-7c44-4b2e-9a01-5f6d8c7b3a10'
-const TURN_ID = '0c5b1e77-2d4a-4f9e-8b63-1a2c3d4e5f60'
-const SOCKET_SID = '7d1f2e3a-4b5c-4d6e-8f90-1a2b3c4d5e6f'
-const HOST_ACTOR = 'agent:comfy'
-const DOC_PROTOCOL_VERSION = 1
-const PANEL_MOUNT_TIMEOUT = 30_000
-const SUBSCRIBE_TIMEOUT = 15_000
-
-const OPEN_AGENT_LABEL = enMessages.agent.askComfyAgent
-const SEND_LABEL = enMessages.agent.send
-const STOP_LABEL = enMessages.agent.stop
-
-interface DocFrame {
-  type: 'doc_subscribed' | 'doc_update'
-  data: Record<string, unknown>
-}
-
-function toBase64(bytes: Uint8Array): string {
-  return Buffer.from(bytes).toString('base64')
-}
-
-function fromBase64(value: string): Uint8Array {
-  return new Uint8Array(Buffer.from(value, 'base64'))
-}
-
-function decodeJson(raw: string): unknown {
-  try {
-    return JSON.parse(raw)
-  } catch {
-    return undefined
-  }
-}
-
-/**
- * The doc host's side of the CRDT channel, in-process: the seed workflow is
- * minted with the same pinned applier the real host runs, each `graph_ops`
- * entry is applied to it, and the resulting Yjs delta is what the follower
- * receives as a `doc_update` frame.
- */
-class HostDoc {
-  private readonly doc: Y.Doc
-  private seq = 1
-
-  constructor(
-    private readonly workflowId: string,
-    seed: WorkflowJSON,
-    private readonly catalog: WidgetCatalog
-  ) {
-    this.doc = mint(seed, catalog)
-  }
-
-  subscribed(): DocFrame {
-    return {
-      type: 'doc_subscribed',
-      data: {
-        v: DOC_PROTOCOL_VERSION,
-        workflow_id: this.workflowId,
-        ok: true,
-        seq: this.seq
-      }
-    }
-  }
-
-  catchUp(stateVectorB64: string): DocFrame {
-    const update = Y.encodeStateAsUpdate(this.doc, fromBase64(stateVectorB64))
-    return this.updateFrame(update, HOST_ACTOR, [])
-  }
-
-  apply(operations: GraphOperation[]): DocFrame {
-    const before = Y.encodeStateVector(this.doc)
-    const ops = mintWireOps(operations, {
-      actor: HOST_ACTOR,
-      baseVersion: this.seq
-    })
-    const result = applyOps(this.doc, ops, this.catalog)
-    const rejected = result.outcomes.filter((o) => o.outcome !== 'applied')
-    if (rejected.length > 0)
-      throw new Error(
-        `conversation graph_ops did not apply: ${JSON.stringify(rejected)}`
-      )
-    this.seq += 1
-    return this.updateFrame(
-      Y.encodeStateAsUpdate(this.doc, before),
-      HOST_ACTOR,
-      ops.map((op) => op.op_id)
-    )
-  }
-
-  private updateFrame(
-    update: Uint8Array,
-    actor: string,
-    opIds: string[]
-  ): DocFrame {
-    return {
-      type: 'doc_update',
-      data: {
-        v: DOC_PROTOCOL_VERSION,
-        workflow_id: this.workflowId,
-        seq: this.seq,
-        update_b64: toBase64(update),
-        actor,
-        op_ids: opIds
-      }
-    }
-  }
-}
-
-interface GraphNodeSnapshot {
-  id: string
-  title: string
-  /** Connected state of each rendered input slot, in slot order. */
-  inputs: boolean[]
-  /** Connected state of each rendered output slot, in slot order. */
-  outputs: boolean[]
-}
-
-/** Subgraph-scoped ids are strings, and `NaN` leaves the order undefined. */
-function byNodeId(a: GraphNodeSnapshot, b: GraphNodeSnapshot): number {
-  const left = Number(a.id)
-  const right = Number(b.id)
-  if (Number.isNaN(left) || Number.isNaN(right)) return a.id.localeCompare(b.id)
-  return left - right
-}
-
-/**
- * Drives one recorded user prompt -> agent response exchange through the
- * real agent panel: the composer posts the prompt, the mocked backend acks
- * it, and the recorded response is replayed over the routed `/ws` socket as
- * chat frames plus doc-host `doc_update` frames the CRDT follower applies to
- * the canvas.
- */
-class AgentConversationHarness {
-  readonly postedMessages: string[] = []
-  readonly panel: Locator
-  readonly ack: AgentTurnAccepted
-
-  private readonly host: HostDoc
-  private socket: WebSocketRoute | null = null
-  private resolveSubscribed: (() => void) | null = null
-  private readonly subscribed = new Promise<void>((resolve) => {
-    this.resolveSubscribed = resolve
-  })
-
-  constructor(
-    private readonly page: Page,
-    readonly conversation: AgentConversation
-  ) {
-    const { workflow } = conversation
-    this.host = new HostDoc(workflow.id, workflow.seed, workflow.catalog)
-    this.ack = {
-      thread_id: THREAD_ID,
-      message_id: TURN_ID,
-      workflow_id: workflow.id
-    }
-    this.panel = page.locator('#agent-panel-root')
-  }
-
-  async boot(agentFlag: boolean): Promise<void> {
-    await this.mockAgentApi()
-    // The app may reopen `/ws` (the connect awaits an auth token), and the
-    // follower only re-drives a pending subscribe on a `status` frame, which
-    // the real server sends on every connect. So every routed socket becomes
-    // the live one and gets that frame.
-    await this.page.routeWebSocket(/\/ws/, (socket) => {
-      this.socket = socket
-      socket.onMessage((raw) => this.onClientFrame(raw))
-      socket.send(
-        JSON.stringify({
-          type: 'status',
-          data: {
-            status: { exec_info: { queue_remaining: 0 } },
-            sid: SOCKET_SID
-          }
-        })
-      )
-    })
-    await bootAgentApp(this.page, agentFlag, {
-      // Follower edits land in the ECS stores, which only the Vue node
-      // renderer projects onto the canvas.
-      settings: { 'Comfy.VueNodes.Enabled': true }
-    })
-
-    await this.page.getByRole('button', { name: OPEN_AGENT_LABEL }).click()
-    await expect(this.panel).toBeVisible({ timeout: PANEL_MOUNT_TIMEOUT })
-  }
-
-  async sendPrompt(): Promise<void> {
-    const { content } = this.conversation.request
-    const alreadyPosted = this.postedMessages.length
-    const composer = this.panel.getByRole('textbox', {
-      name: /^Describe ideas/
-    })
-    await composer.fill(content)
-    await this.panel.getByRole('button', { name: SEND_LABEL }).click()
-    await expect
-      .poll(() => this.postedMessages.length)
-      .toBeGreaterThan(alreadyPosted)
-    const posted: unknown = JSON.parse(this.postedMessages[alreadyPosted])
-    expect(posted).toMatchObject({ content })
-  }
-
-  async replayResponse(): Promise<void> {
-    for (const entry of this.conversation.response) {
-      if (entry.kind === 'event') {
-        this.send(this.stampTurn(entry.event))
-        continue
-      }
-      await this.waitForSubscribe()
-      this.send(this.host.apply(entry.ops))
-    }
-  }
-
-  async waitForTurnComplete(): Promise<void> {
-    await expect(
-      this.panel.getByRole('button', { name: SEND_LABEL })
-    ).toBeVisible()
-    await expect(
-      this.panel.getByRole('button', { name: STOP_LABEL })
-    ).toHaveCount(0)
-  }
-
-  /**
-   * Every node rendered on the canvas, with its slot connectivity. The tab
-   * the agent opens is expected to be blank (AgentPanelRoot mints it from
-   * `blankGraph`), so anything beyond the host doc's own nodes is the
-   * regression this snapshot exists to catch and must not be filtered away.
-   */
-  async graphSnapshot(): Promise<GraphNodeSnapshot[]> {
-    const snapshot = await this.page
-      .locator('[data-node-id]')
-      .evaluateAll((nodes) =>
-        nodes.map((node) => {
-          const connected = (selector: string) =>
-            Array.from(node.querySelectorAll(selector)).map((slot) =>
-              slot.classList.contains('lg-slot--connected')
-            )
-          const title = node.querySelector('[data-testid="node-title"]')
-          return {
-            id: node.getAttribute('data-node-id') ?? '',
-            title: title === null ? '' : title.textContent.trim(),
-            inputs: connected('.lg-slot--input'),
-            outputs: connected('.lg-slot--output')
-          }
-        })
-      )
-    return snapshot.sort(byNodeId)
-  }
-
-  private async mockAgentApi(): Promise<void> {
-    const { page } = this
-    await page.route('**/api/agent/threads', (route) =>
-      route.fulfill(jsonRoute({ threads: [] }))
-    )
-    await page.route('**/api/agent/threads/*/messages', (route) => {
-      const request = route.request()
-      if (request.method() === 'POST') {
-        this.postedMessages.push(request.postData() ?? '')
-        return route.fulfill({
-          status: 202,
-          contentType: 'application/json',
-          body: JSON.stringify(this.ack)
-        })
-      }
-      const history: AgentMessages = []
-      return route.fulfill(jsonRoute(history))
-    })
-    await page.route('**/api/agent/threads/*/messages/*/cancel', (route) => {
-      const cancelled: AgentCancelAccepted = { status: 'cancelling' }
-      return route.fulfill(jsonRoute(cancelled))
-    })
-    await page.route('**/api/workflows**', (route) =>
-      route.fulfill(
-        jsonRoute({
-          data: [],
-          pagination: { has_more: false, limit: 100, offset: 0, total: 0 }
-        })
-      )
-    )
-  }
-
-  private stampTurn(event: RecordedWsEvent): AgentWsEvent {
-    const stamped = {
-      type: event.type,
-      data: { ...event.data, message_id: TURN_ID, thread_id: THREAD_ID }
-    }
-    const parsed = parseAgentWsEvent(stamped)
-    if (!parsed.success)
-      throw new Error(
-        `recorded ${event.type} frame is not a valid agent event: ${parsed.error.message}`
-      )
-    return parsed.data
-  }
-
-  private send(frame: AgentWsEvent | DocFrame): void {
-    if (!this.socket) throw new Error('the app has not opened /ws yet')
-    this.socket.send(JSON.stringify(frame))
-  }
-
-  private onClientFrame(raw: string | Buffer): void {
-    // The route intercepts the whole `/ws` channel, so a non-JSON client
-    // frame must not throw inside Playwright's websocket handler.
-    const frame = decodeJson(raw.toString())
-    if (typeof frame !== 'object' || frame === null) return
-    const { type, data } = frame as { type?: unknown; data?: unknown }
-    if (type !== 'doc_subscribe' || typeof data !== 'object' || data === null)
-      return
-    const { workflow_id, state_vector_b64 } = data as {
-      workflow_id?: unknown
-      state_vector_b64?: unknown
-    }
-    if (
-      workflow_id !== this.conversation.workflow.id ||
-      typeof state_vector_b64 !== 'string'
-    )
-      return
-    this.send(this.host.subscribed())
-    this.send(this.host.catchUp(state_vector_b64))
-    this.resolveSubscribed?.()
-  }
-
-  private async waitForSubscribe(): Promise<void> {
-    let timer: ReturnType<typeof setTimeout> | undefined
-    const timeout = new Promise<never>((_, reject) => {
-      timer = setTimeout(
-        () =>
-          reject(
-            new Error(
-              'the follower never subscribed to the conversation workflow; graph_ops need an agent_active_tab (or a bound tab) first'
-            )
-          ),
-        SUBSCRIBE_TIMEOUT
-      )
-    })
-    try {
-      await Promise.race([this.subscribed, timeout])
-    } finally {
-      clearTimeout(timer)
-    }
-  }
-}
+import { AgentConversationHarness } from '@e2e/fixtures/helpers/AgentConversationHarness'
 
 interface ConversationFixtures {
   /** Case id of the conversation under `fixtures/data/agent/conversations`. */
